goodbye, $.parseFromISO
the vast majority of calls were only to access:

 * the parsed-then-fudged value; this can be accomplished with just
   fudgeDateForProfileTimezone (given an ISO string it will parse it,
   format it relative to the profile timezone (without tz info), then parse
   it relative to the browser timezone. unfudging this result will give
   you back a time accurately representing the original string)

 * one or more formattings of the parsed value; use the appropriate
   formatter (if multiple are used, parse with tz.parse() first for
   efficiency)

one notable exception was in the gradebook (both regular and
screenreader). it was using parseFromISO to parse the ISO string
received from the server, but then only sometimes treating that as a
object while other times assuming (incorrectly) that it had a Date
object or timestamp. we make it always be a Date object, and fix the
places that assumed it needed to be a parseFromISO result object.

Change-Id: I86c077c046e74760a538849ce58f2952ada1fb99
Reviewed-on: https://gerrit.instructure.com/32110
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Anthus Williams <awilliams@instructure.com>
Product-Review: Jacob Fugal <jacob@instructure.com>
QA-Review: Jacob Fugal <jacob@instructure.com>