Ensure overdue quiz submission are graded
In cases where a student taking a quiz that has a time-limit closes the
window without submitting, their submission would never be graded until
they re-visit the session, leaving teachers with an ungraded submission
that can not be moderated.

This patchset modifies the QuizzesController so that it queues a job the
moment a student starts taking a quiz with a time-limit to be run a few
seconds after the submission becomes overdue.

closes CNVS-2337

Test steps:

  - Create a quiz with a time-limit (1 minute will do)
  - In a tab, as a teacher, go to the Moderate Quiz view
  - In another tab, as a student, start taking the quiz
  - Choose some correct answer so you get some sensible score later
  - Take a look at the teacher tab, it should display a countdown timer
    in the respective student entry with no score (yet)
  - Close the student tab
  - Wait-out the timer, keep an eye on the teacher's tab
  - 6-8 seconds after the time goes out, you should now see the
    "Time Up!" message disappear, and the score column be populated

Optionally, you can repeat the steps with another quiz attempt, and the
score should again reflect the latest/highest submission as per the quiz
settings.

Change-Id: I6ec7ef52e35879b4f8c86147b48fa2758f7b88da
Reviewed-on: https://gerrit.instructure.com/25400
Reviewed-by: Derek DeVries <ddevries@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Myller de Araujo <myller@instructure.com>
Product-Review: Derek DeVries <ddevries@instructure.com>