Correct answer selection keyboard a11y
A CSS directive that renders the correct answer arrow already existed
but was not valid because it was trying to manipulate a parent (the
.answer_box element) which actually renders the background-image.

This patch necessarily modifies the rendering of the image to happen in
the innermost child (the img.answer_image) element so it can be
manipulated by the enclosing a.select_answer_link (on :focus) as well as
the enclosing .answer box (on .hover)

On the JS side, flagging an answer as the correct one will now remove
the .hover state from the previously hovered element. This is to address
the confusion this caused when flagging using a keyboard.

fixes CNVS-6547

-- Test plan

* create a quiz
* add a question with multiple answers
  - hover over the answers, the "partial answer" (faded arrow) should appear
  - choose some input field
  - press TAB until you reach the next answer's "correct answer" arrow:
    * the image should be rendered with a thin outline
  - press RETURN
    * the image should no longer be faded
    * answer should be highlighted as the correct one
    * answer under your mouse cursor should no longer be highlighted as
    if hovered over

Change-Id: Iefe71e849a134a5757257c7f243cfa217a1ce67c
Reviewed-on: https://gerrit.instructure.com/25755
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Derek DeVries <ddevries@instructure.com>
QA-Review: Myller de Araujo <myller@instructure.com>
Product-Review: Derek DeVries <ddevries@instructure.com>