align due_date text on the module index page
The due date text was overflowing over the completion requirement text.
This commit fixes it so that doesn't happen anymore.

fixes CNVS-9643

Test Plan
With Draft State Enabled
1. Go to the modules index page and add an assignment.
2. To that assignment, add a due date to that assignment.
3. Add a requirement for that assignment in that module. The best one
to test is "Must contribute to the content of the page" because it's
has the logest description and best illustrates the problem.
4. View the assignment on the modules index page and observe that the
text for requirements to complete the module item and it's due date do
NOT overflow.

Change-Id: Ia7234e411ab0832f2a7681c8d4cf7c6b9dd33799
Reviewed-on: https://gerrit.instructure.com/26937
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Matt Fairbourn <mfairbourn@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
Product-Review: Jeremy Stanley <jeremy@instructure.com>