"Only Teachers" means "Only users with manage_wiki rights"
Make it so that when a wiki page is set to
 [Only Teachers] can edit this page
it really means that only users who have manage_wiki rights
can edit the page. Otherwise, when a teacher role is denied
manage_wiki rights (and there are a handful of schools that
do this), teachers can still edit wiki pages, and it's
impossible to take that right away.

Also make "Teachers and students" give rights to all course
admins (Teachers, TAs, and Designers) in addition to
Students, because it's weird to give edit rights to
students but not course admins.

fixes CNVS-1794

test plan:
 - Ensure that teachers, TAs, and designers (who are not
   account or site administrators) can still create and
   edit wiki pages when default permissions are left alone
 - Ensure that custom teacher/TA/designer roles with
  "manage wiki" removed:
   - Cannot create wiki pages
   - Cannot delete wiki pages
   - Cannot edit wiki pages that are set to "Only teachers"
   - Can edit wiki pages that are set to "Teachers and Students"
   - Can edit wiki pages that are set to "Anyone"

Change-Id: Icb24d4f86833315b94884eadee76131fe41f58b3
Reviewed-on: https://gerrit.instructure.com/27449
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Nathan Rogowski <nathan@instructure.com>
Reviewed-by: Bracken Mosbacker <bracken@instructure.com>
Product-Review: Bracken Mosbacker <bracken@instructure.com>