Add integration_id to SIS models
An integration_id can be used in a way similar to sis_source_id in
that it documents a 3rd party identifier for a particular object. The
reason we use an integration_id instead of a sis_source_id for some
SIS integrations is that there is BOTH a sis_source_id and another
id that is used (as in the case of SIF) to ferry information between
systems.

Models that have an integration_id now include:
- pseudonyms
- course_sections
- courses
- enrollment_terms
- accounts

Test Plan:
- make sure the above tables have an integration_id column after "rake db:migrate"

Fixes SIS-107

Change-Id: I80a03d7384fc6ff243436bdf18798ece5059acb7
Reviewed-on: https://gerrit.instructure.com/27593
Reviewed-by: Cody Cutrer <cody@instructure.com>
Reviewed-by: Ken Romney <kromney@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Brian Palmer <brianp@instructure.com>
Product-Review: Duane Johnson <duane@instructure.com>
QA-Review: Duane Johnson <duane@instructure.com>