QS API - Respect the quiz's LDB requirement
Attempts to create or complete a quiz submission via the API will now be
rejected if the quiz requires LDB and the client is not identified to be
using one.

closes CNVS-10220

A note to reviewers: the filter is at the controller layer instead of
the service one because the LDB plugin requires a controller to validate
the authorization, and it seemed silly to pass the controller to either
the service or the QuizParticipant.

TEST PLAN
---- ----

  - create a quiz that requires the LDB (you may have to install/enable
    the LDB plugin for this to work)
  - try to create and complete the quiz submission via the API from a
    regular browser (like Chrome with Postman) or via cURL (see below)
    * your request should be rejected with a 403 telling you the quiz
      requires LDB

cURL command:

curl -X POST \
  -d '{}' \
  -H 'Accept: application/json' \
  -H 'Authorization: Bearer YOUR_API_TOKEN_HERE' \
  http://localhost/api/v1/courses/:course_id/quizzes/:quiz_id/submissions

Change-Id: Ic9d3561cd21288ff51d1946854f416c92b4a9f57
Reviewed-on: https://gerrit.instructure.com/28021
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Myller de Araujo <myller@instructure.com>
Reviewed-by: Stanley Stuart <stanley@instructure.com>
Product-Review: Ahmad Amireh <ahmad@instructure.com>