don't submit to scribd if it has a scribd_doc
refs CNVS-10818

this will check that it is a root_attachment
then check to see if it has a scribd_doc
then make sure the status of the doc is good.
when documents are good we don't re-send them to
scribd. when documents are bad we re-send them.

test plan
 - scribadable documents should still work

Change-Id: I676cc282f0e7c824adce0048c706de71b39cb358
Reviewed-on: https://gerrit.instructure.com/28423
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Rob Orton <rob@instructure.com>
Product-Review: Rob Orton <rob@instructure.com>
QA-Review: Rob Orton <rob@instructure.com>