fix issue where speedgrader sidebar doesn't load in quiz submission attempts
fixes CNVS-11994

test plan:

the problem on this one comes up when we are trying to view submission versions
that were created before the namespace change. So we have to massage the data in
the console to reproduce what's going on here, or test it with data that was
created before the namespace changes.

- as a teacher
  - create a quiz that allows multiple attempts

- as a student
  - take the quiz two times

- in the console
  Version.update_all("versionable_type='QuizSubmission'", "versionable_type='Quizzes::QuizSubmission'")

- as a teacher
  - visit speedgrader for the student's submission
  - you should be able to switch between quiz attempts in the sidebar

Change-Id: I6e16abd91dcaa8d42bf66bcbacab0cce6eaed9e6
Reviewed-on: https://gerrit.instructure.com/32383
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Myller de Araujo <myller@instructure.com>
Reviewed-by: Cameron Matheson <cameron@instructure.com>
Product-Review: Derek DeVries <ddevries@instructure.com>