fix quiz grading when a quiz with a time limit is exited prematurely
fixes CNVS-12074
If a student takes a quiz with a time limit, and quits the quiz
without explicitly submitting it, the quiz will get stuck in a state
where it won't be automatically graded. This commit fixes the #needs_grading?
method on QuizSubmission to prevent this.

Test plan
- As a teacher, create a quiz with a time limit
- As a student, take the quiz, but close the browser before the quiz
  auto submits.
- As the teacher, navigate to the moderate page, and click on the
   student's submission
- The student's submission should show up and be graded.

Change-Id: I161937365f01d0f87e2cb56c8579c060337a8dd9
Reviewed-on: https://gerrit.instructure.com/32538
Reviewed-by: Jason Madsen <jmadsen@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Caleb Guanzon <cguanzon@instructure.com>
Product-Review: Josh Simpson <jsimpson@instructure.com>