Remove assets.yml refs to files that don't exist
there were a few files that we were still pointing
to in our assets.yml css bundles that the actual
css file does not even exist. when that happens,
jammit does not throw an error or anything but
just prints a warning to stdout.  we probably
deleted the sass/css file long ago but forgot to
remove the reference to it in assets.yml

the only thing I can think of that would be dangerous
here is if one of our plugins depended on one of
these bundles existing so it could override it,
so I'll put some of our plugin experts on review

test plan:
verify rake canvas:compile_assets works and doesn't
print Jammit Warning: No assets match ... warnings

Change-Id: I0f356861159e790302fe14d7e74c3e9172c177f5
Reviewed-on: https://gerrit.instructure.com/34010
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Jacob Fugal <jacob@instructure.com>
Reviewed-by: Jon Jensen <jon@instructure.com>
QA-Review: Nathan Rogowski <nathan@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>