accessiblity touch ups when copying a course
fixes CNVS-12188

When you are migrating a course and using voice over in safari it was
really confusing because when you selected 'copy course' and did a
course search or a course select, the results would be read we everything else on the
page. This fixes that.

Test Plan
As a teacher or someone who can run a migration
Given you are using safari with voice over
When you import content into a course
And you select the content type of 'Copy a Canvas Course'
And you search for a course by its name by using the field
Then voice over should tell you there are results
And when you use the arrow keys to select one of those results
Then only the selected result should be read by voice over
And the all of the form items should not be read
----------------------------------------------------
As a teacher or someone who can run a migration
Given you are using safari with voice over
When you import content into a course
And you select the content type of 'Copy a Canvas Course'
And you select a course in the course select dropdown
And when you use the arrow keys to select one of those results
Then only the selected result should be read by voice over
And the all of the form items should not be read

Change-Id: Id2c265d900ee9676094e221fa9dd38e8a61a843a
Reviewed-on: https://gerrit.instructure.com/35240
Tested-by: Jenkins <jenkins@instructure.com>
Product-Review: Hilary Scharton <hilary@instructure.com>
QA-Review: Clare Strong <clare@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>