Better formatting for emails and add avatar for conversations
Fixes: CNVS-12970

We needed an avatar and signature for notifications
sent out via conversations. Also reformatted all the
notifications to take on a cleaner and more easy to
read format.

The majority of testing can be done on the conversation
notifications. I was using letter opener to do this.
I have included letter opener and the relevant setup
stuff in this commit so that it can be tested.

Test-Plan:
-Make sure that your organization has allowed avatars
-Give a student or a teacher an avatar
-Send a conversation message to another user from
  the user that has an avatar
-If you have letter_opener installed, the email
  should be intercepted and you'll see its formatting.
  Otherwise, use whatever method you'd like to view
  the email.
-Verify that everything looks alright and we have
  an avatar and good formatting.

-STEP 2-
-Disable avatars
-Send another email and verify that the avatar isn't
  showing up

Change-Id: I876dc119c5daefa10811c75c08b315722a50f63d
Reviewed-on: https://gerrit.instructure.com/35144
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Ethan Vizitei <evizitei@instructure.com>
QA-Review: Trevor deHaan <tdehaan@instructure.com>
Product-Review: Drew Bowman <dbowman@instructure.com>