wait a tick to display quiz submission time
when a student has a previous submission and then starts a new one, it
was possible for the previous quiz submission time to temporarily be
displayed in the moderate table. so wait a tick until we display
the quiz submission time counter to verify that all the quiz submission
properties and bindings have had a chance to fire.

closes CNVS-13342

test plan:
- create a quiz with:
  - time limit
  - or pending due date
  - and multiple attempts
- as a student take the quiz, but don't submit it
- as a instructor visit the moderate page
  - verify that submission time is running
- as student, submit the quiz
- as instructor, click the refresh icon and verify that submission time
  stops running
- as student take the quiz again
- as instructor click the refresh icon again and verify that previous
  running time doesn't show temporarily.

Change-Id: Ie501e7ffa31fe4b4959eeb783cfaf868e0f0aa64
Reviewed-on: https://gerrit.instructure.com/35618
Reviewed-by: Josh Simpson <jsimpson@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Caleb Guanzon <cguanzon@instructure.com>
Product-Review: Jason Madsen <jmadsen@instructure.com>