Accessibility: Alert message added for message delete
Fixes: CNVS-12802

An alert message was needed to inform screen readers
that the message was deleted. I decided to go with
both audio and visual alert, but this can be easily
changed to just audio if it's decided that the
visual notification in unnecessary.

Test-Plan:
Using a screenreader
1.) Navigate to conversations
2.) Delete a message
3.) Verify that an audible and visual alert exists

Change-Id: I68a54074df7eff2b9d04f3d99d6617a77208488c
Reviewed-on: https://gerrit.instructure.com/35883
Reviewed-by: Aaron Cannon <acannon@instructure.com>
Tested-by: Jenkins <jenkins@instructure.com>
QA-Review: Steven Shepherd <sshepherd@instructure.com>
Product-Review: Jayce Higgins <jhiggins@instructure.com>
Reviewed-by: Drew Bowman <dbowman@instructure.com>