disable custom override css while working on k12 and new_styles
fixes: CNVS-14438

at least while we are working on the new style and
k12 specific looks, let's not load customer account
css overrides.  because we are doing major changes
to the header and nav in these new features, we
don't wan't to be pulling in the css files that
people were loading before since they would probably
be all broken in undetermined ways now. However,
because all this work is behind a feature flag,
and these feature flags are marked as "in development"
and "beta", we are not removing anything for any
customers.

repeated:
This does not affect anyone that does not have
either the k12 or new_styles feature flags enabled.
so none of our current customers in production.

As we move closer to having the new_syles and k12
styles out of beta we'll work on a better story
for branding canvas to a school in a easier, more
manageable way. This may involve a GUI interface
to pick your colors instead of having to upload a
css file that overwrites things in an unsupported
and undocumented way. We don't want people to have
to worry about a new release breaking their branding.

test plan:

with k12 or new_styles feature flag turned on,
make sure the account's custom css file is not loaded

with neither the k12 or new_styles feature flag turned on
make sure that nothing changes, e.g. if they had custom
css it should still be loaded as before.

Change-Id: I4ec4bbac55d6e913d19fddface1232baf6d7ff89
Reviewed-on: https://gerrit.instructure.com/38294
Tested-by: Jenkins <jenkins@instructure.com>
Reviewed-by: Jon Willesen <jonw@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>